+++ SPI_device.cpp
@@ -9,7 +9,7 @@ SPI_device::~SPI_device()
 {
 }
 
-void SPI_device::txrx( uint8_t *data, uint16_t size )
+void SPI_device::txrx( uint8_t *data, int size )
 {
 	digitalWrite( SS, LOW );
 	SPI.transfer( data, size );
